Update the sample_config
This commit changes the paths in sample_config so that the directories
are consolidated in $XDG_DATA_HOME instead of being split all over the
place (like my files were organized during testing). This might make it
easier for people using this code to ensure that all of their files are
stored in close proximity to each other.