Update code to work with single database
The initial version of the code was working with two database files - a
BibTeX file to store the references and a YAML file to store additional
attributes relating to the various entries. This commit implements a
complete overhaul of this system so that only a single YAML file is now
needed to store all the information, hopefully making the storage and
searching easier.