Add addfile command
This commit introduces the command to be able to add files. In addition,
it adds several utilities to be able to accomplish this (some of which
can be used to query crossref.org). Additionally, some of the
functionality of addentry has been changed to make it play a little
nicer with addfile. Further, the key format has been changed from Lf-Y
to fL-Y.