Preview a file while adding it
When adding a file, it would be easier to verify / modify the entry
associated with it if the file was opened. This commit opens the file
using whatever program is set in the $READER environment variable. If
this is not set, it can also be set in the config file.