Update editnote to use template and metadata
The editnote command can now use a template. Also, the metadata for the
record is added to the file. The user is also able to update the
metadata which will be stripped from the file when the note is saved and
then cause an update of the appropriate record in the database.