Allow creation of manual entries
This allows to create entries of type "misc" which can then be used to
cite web pages, lecture slides, personal notes, videos, and other such
sources that do not fit into any of the other BibTeX categories. The
main purpose of this is to allow for notes from these sources to also be
stored within clarence's directories.