Add fuzzy search
Up until now, if you wanted to search for something in the database, you
could provide a list of strings on the command line and have clarence
search the database and return all records that match. I got really
tired of reading through this [very long] list of records to find the
record that I wanted. So I decided to use fzf (a fuzzy finder) to make
the process of searching a little bit easier. This commit introduces
this feature.

This feature still needs to be tied in with the readpdf and editnote
commands. But even without doing that, it still feels like a significant
improvement.