Add summary file feature
We already have the capability to write notes on specific entries in the
database. However, I wanted a way to synthesize and record information
that was related (for example, about a project or a topic). That's what
summary files are for.

This commit adds in some very basic features for creating summary files
and searching through them. The idea for this code was taken from stnote
(a terminal notes app I wrote).