Add feature to insert citations in vim
This commit makes a small modification to the fuzzy search feature so
that it can be used from within vim to insert citations. It also
provides a snippet that can be inserted into a .vimrc to enable this
functionality.

An earlier solution that was tried turned off shelltemp in vim (set
noshelltemp). This did not work consistently - it would freeze up every
once in a while and/or not format the citations correctly. So I decided
that it was better to just use the tempfile method (which is an idea
that I got from
https://stackoverflow.com/questions/24299155/capture-the-output-of-an-interactive-script-in-vim).