Allow fuzzy searching in previous commands
Commands like editnote and readpdf which have existed since before fuzzy
search features were introduced didn't have a way to exploit the
convenience of the fuzzy search feature. The only way to do it was to
pass in a key for a resource (which could be done by running a fuzzy
search and piping the resulting key into the appropriate command) which
was very annoying.

This commit allows them to invoke fuzzy search internally if they are
run without any arguments. The previous functionality of providing a key
argument and the command line completion should still work as they
normally do.