Show resource type when querying crossref
When querying crossref, some resources could not be added because
clarence did not know how to add a resource of that type. Sometimes
there were also records on crossref that had similar metadata stored
with different types. This commit prints out the type information when
printing out the abbreviated crossref results so that the user knows
what type the resource is.