Cut after getting finding doi in pdf2doi
In cases when the DOI finding statement would not find a DOI number, the
cut command would still run successfully causing the script to produce
garbage output instead of exiting and saying that it didn't find a DOI
number. This commit moves the cut command to a new statement so that the
script fails first if no DOI number is found.