Add option to remove references in doi2bibjson
In some cases, running doi2bibjson was producing a lot of output which
would cause addentry (and hence addfile) to fail. This was because when
the reference information is included in the JSON output, the command
line argument piped into addentry (the JSON) is too long.

This commit adds in an option to allow the user to remove the reference
information from the JSON output returned from crossref. Hopefully this
implementation allows us to reduce the length of the JSON returned by
crossref to levels acceptable for passing in as command line arguments.
Also, perhaps someday we can just download the JSON into a tmpfile and
read that into the addentry script instead of passing the entire JSON
string in as a command line argument.

Note that I have not tested this extensively at all (read: it was tested
on one example that was failing when I tried to run addfile on it). I
don't know whether or not this will work or cause old functionality to
break. Hopefully only good things happen!