Remove references from bibjson during addfile
Sometimes the BibJSON returned from crossref is too long to be passed in
as a command line argument to addentry because it has the list of
references contained in it. This commit removes this list of references
so that the BibJSON is shorter in length and can be passed in as a
command line argument (allowing addfile and addentry to function without
having to make larger changes).