Clean control characters in crossref json response
I don't know if the API is returned results in a different format now or
if this is a weird edge case that never got tested, but I started
observing new lines coming back in the response that were messing up how
jq was interpreting the results string. This commit removes the new
lines so that the response is all on one line and can be parsed by jq
appropriately.