Increase POSIX compliance
Testing clarence with dash highlighted several areas in the code that
were not POSIX compliant and so did not work with the dash shell. This
commit is an attempt to fix most of those issues. I have not been able
to test the entirety of the code to ensure that it runs with dash. But,
hopefully, as I begin to use it, I'll be able to achieve more coverage.