Add readsummary feature
This extends the summaries feature. So far you could only read the
summaries by opening them up in a text editor. With this commit, you can
also read them as PDF files - the markdown file is passed through pandoc
to create a PDF file which is displayed on the fly.