Allow files and entries to be added
This is the first commit with the first feature - to add files and
entries. There are still some unhandled situations in this version but
I'll add them in as time goes on. It's been an interesting journey
rewriting this portion of the code so many times to get to this stage.