Ignore results in main for the time being
At this time, I don't have any logic for what I want to do with the
results I get from calling the methods on library. Eventually, I'll want
to print this stuff to the screen. But the library methods will end up
returning different types (as I learned when I started implementing the
searching functionality). This is not compatible with the match
statement. I think what I want to do is implement some convenience
methods to format the results as text and print them. So I'll just come
back to this piece later.